feat: make bridge.Refresh return if package list changed
