+++ bridge/actions.go
@@ -2,7 +2,6 @@ package bridge
 
 import (
 	"fmt"
-	"log"
 	"strings"
 
 	"github.com/lavafroth/droidrunco/app"
@@ -10,16 +9,16 @@ import (
 )
 
 var Cache app.Apps
-var Updated bool
 
-func Refresh() {
+// Refresh returns true if the refresh resulted in a changed set of packages
+func Refresh() (bool, error) {
 	out, err := device.RunCommand("pm list packages -f")
 	if err != nil {
-		log.Fatalf("failed to fetch list of packages: %q", err)
+		return false, fmt.Errorf("failed to fetch list of packages: %q", err)
 	}
 	out = strings.Trim(out, "\n\t ")
 	var fresh app.Apps = map[string]*app.App{}
-	gotFreshPackages := false
+	changed := false
 
 	for _, line := range strings.Split(out, "\n") {
 		_, line, _ := strings.Cut(line, ":")
@@ -30,7 +29,7 @@ func Refresh() {
 		// we don't bother looking up its label name.
 		App, ok := Cache[id]
 		if !ok {
-			gotFreshPackages = true
+			changed = true
 
 			metadata := meta.Meta{}
 
@@ -38,46 +37,41 @@ func Refresh() {
 				metadata = *metadataPtr
 			}
 
-			HasLabel := metadata.Label != ""
-
-			App = &app.App{Id: id, Meta: metadata, Path: path, Enabled: true, HasLabel: HasLabel}
+			App = &app.App{Id: id, Meta: metadata, Path: path, Enabled: true}
 		}
 		fresh[id] = App
 	}
 
-	if !gotFreshPackages {
-		return
+	// has any app been disabled? Present in old map, absent in new.
+	for id, app := range Cache {
+		if _, inNewMap := fresh[id]; !inNewMap {
+			changed = true
+			app.Enabled = false
+			fresh[id] = app
+		}
+	}
+
+	if !changed {
+		return false, nil
 	}
 
 	out, err = device.RunCommand("CLASSPATH=/data/local/tmp/extractor.dex app_process / Main")
 	if err != nil {
-		log.Fatalf("failed to fetch list of packages: %q", err)
+		return false, fmt.Errorf("failed to extract package labels: %q", err)
 	}
 	out = strings.Trim(out, "\n\t ")
 	for _, line := range strings.Split(out, "\n") {
 		_, idAndLabel, _ := strings.Cut(line, " ")
 		id, label, _ := strings.Cut(idAndLabel, " ")
 
-		// If we can already find the same package in the old list,
-		// we don't bother looking up its label name.
 		App, ok := fresh[id]
-		if ok {
+		if ok && App.Label == "" {
 			App.SetLabel(label)
 		}
 	}
 
-	Updated = gotFreshPackages
-
-	for id, app := range Cache {
-		// The app was previously enabled
-		// but is no more in the new list.
-		if fresh[id] == nil {
-			// App has been disabled.
-			app.Enabled = false
-			fresh[id] = app
-		}
-	}
 	Cache = fresh
+	return true, nil
 }
 
 func Toggle(App *app.App) string {
