+++ bridge/bridge.go
@@ -1,12 +1,20 @@
 package bridge
 
 import (
-	"log"
-	// "time"
-
 	"github.com/lavafroth/droidrunco/meta"
 	adb "github.com/zach-klippenstein/goadb"
+
+	_ "embed"
+	"fmt"
+	"log"
+	"time"
 )
+
+// Thank you Irfan Latif
+// https://android.stackexchange.com/questions/90141/obtain-package-name-and-common-name-of-apps-via-adb
+
+//go:embed extractor.dex
+var extractorDex []byte
 
 var device *adb.Device
 var db meta.DB
@@ -38,4 +46,16 @@ func Init() {
 
 func Close() {
 	client.KillServer()
+}
+
+func push(remote string) error {
+	remoteHandle, err := device.OpenWrite(remote, 0o755, time.Now())
+	if err != nil {
+		return fmt.Errorf("failed to open handle with write permissions on file %s: %q", remote, err)
+	}
+	defer remoteHandle.Close()
+	if _, err := remoteHandle.Write(extractorDex); err != nil {
+		return fmt.Errorf("failed to copy data from local file handle to remote file: %q", err)
+	}
+	return nil
 }
